fix conversion errors (with new compile flags)

+++ sources/time/now.cpp
@@ -6,6 +6,6 @@
 
 /* now */
 auto mx::now(void) noexcept -> mx::u64 {
-	return ::mach_absolute_time()
-		* mx::timebase::absolute_to_nano();
+	return static_cast<mx::u64>(::mach_absolute_time()
+							* mx::timebase::ticks_to_ns());
 }
